% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbour_offset.R
\name{local_offset_distance_with_background}
\alias{local_offset_distance_with_background}
\title{Neighbourhood offset}
\usage{
local_offset_distance_with_background(
  mat,
  coords,
  tx,
  tx_x = "x",
  tx_y = "y",
  feature_name = "gene",
  distance = 50,
  nbins = 200,
  cl = 1
)
}
\arguments{
\item{mat}{A count matrix with genes as rows and cells as columns.}

\item{coords}{A dataframe with x, y coordinates of each cell as separate columns.}

\item{tx}{A transcript dataframe with x, y coordinates and qv values.}

\item{tx_x}{Column name for the x coordinates in the transcripts dataframe.}

\item{tx_y}{Column name for the y coordinates in the transcripts dataframe.}

\item{feature_name}{Column name for the gene of each transcript in the transcripts dataframe.}

\item{distance}{The maximum distance to consider for local background estimation.}

\item{nbins}{The number of bins to use for hexagonal binning, used for calculating background transcript contamination.}

\item{cl}{The number of cores to use for parallel processing.}
}
\value{
A matrix of local background counts for each gene in each cell.
}
\description{
This function calculates the local neighbourhood offset together with ambient background for each cell in a count matrix.
}
